+++ test/FakeArduino.cpp
@@ -47,7 +47,7 @@ int32_t arduino_random(int32_t min, int32_t max)
 class LoRaL2;
 
 uint8_t* lora_test_last_sent = 0;
-int lora_test_last_sent_len = 0;
+size_t lora_test_last_sent_len = 0;
 
 bool lora_start(long int band, int spread, int bandwidth, int power,
 		int paboost, int cr4slsh, LoRaL2 *pobserver)
@@ -64,7 +64,7 @@ bool lora_begin_packet()
 	return true;
 }
 
-void lora_finish_packet(const uint8_t* packet, int len)
+void lora_finish_packet(const uint8_t* packet, size_t len)
 {
 	if (lora_test_last_sent) {
 		free(lora_test_last_sent);
